requirements.txt
fire>=0.2.1
matplotlib>=3.1.1
mlflow>=1.2.0
pandas>=0.25.1
plotly>=4.2.1
-e git+https://github.com/huggingface/pytorch-transformers.git@32e1332acf6fd1ad372b81c296d43be441d3b0b1#egg=pytorch_transformers
PyYAML>=5.1.2
scipy>=1.3.1
spacy>=2.2.1
torch>=1.2.0
tqdm>=4.35.0
sklearn>=0.0
